fpga: dfl: check feature type before parse irq info
The feature ID of "Port User Interrupt" and the
"PMCI Subsystem" are identical, 0x12, but one is for FME,
other is for Port. It should check the feature type While
parsing the irq info in parse_feature_irqs().

---
v3: Remove "Fixes" in commit log with Hao's comment, this is a
    extension or a bug fix.

v2: add DFL Feature ID Registry in documentation.

Signed-off-by: Tianfei Zhang <tianfei.zhang@intel.com>