fpga: dfl: fme: adding reserved bits for revision of FME/Port error
There are 2 different register layouts for FME/Port error
registers. The new production like Intel PAC N6000 plans to
deploy the new register format which it is more reasonable.

To distinguish the register layouts,we provide another sysfs
node for revision info, but it is a bad method that using one
sysfs node's value to determine the usage of other sysfs node.

This patch introduces 4 reserved bits (Bit[59:56]) which
are reserved by HW, dfl driver appends the FME/Port error
revision info on those bits for attribution on the readout
value.

Signed-off-by: Tianfei zhang <tianfei.zhang@intel.com>
---
v2:
 - add documentation for this change.
 - fix the reverse xmas tree declaration.