fpga: dfl: add a new driver for image reload manager
Adding a new driver for DFL FPGA image reload manager.
After burn a new FPGA/BMC image, it should stop the
current running FPGA image and reload the new image
from the flash to FPGA. Actually it should stop all of
PFs/VFs and all of DFL subdevices which running on the
old image.

To help manager the image reload, we introduce "dfl-image-reload"
driver, which is on the top of DFL firmware, provide some
operations to implement the reload functionality.

Signed-off-by: Tianfei Zhang <tianfei.zhang@intel.com>