fpga: dfl: add sysfs files for image reload
This patch add 4 sysfs files to dfl image reload manager.
Available_images will returns a list of key words
that may be written into the reload sysfs file. A key
word may be written to reload sysfs file to trigger
a reload of an FPGA, BMC, or firmware image from FLASH
or EEPROM.

Here are the steps of an image reload trigger process.
1. remove all PFs and VFs except the PF0.
2. remove all non-reserved devices of PF0.
3. trigger image reload.
4. disable the link of pci port which will be removed later,
   eg, pci root hub.
5. remove all reserved devices under PF0 and PCI devices
   under the pci port.
6. Wait for FPGA/BMC reload done, e.g. 10s.
7. enable the link of pci port.
8. rescan the PCI bus.

Signed-off-by: Tianfei Zhang <tianfei.zhang@intel.com>