PCI: hotplug: expose APIs from pciehp driver
When a PCIe-based FPGA card is reprogrammed, it temporarily
disappears from the PCIe bus. This needs to be managed by
disabling the link to avoid PCIe errors while the device is
not present. Also, re-enabling the link and rescan the PCI
devices must be performed after loading the new images. Export
functions from pciehp driver necessary for removing and
reconfiguring the PCI devices below a PCI hotplug bridge.

Signed-off-by: Tianfei Zhang <tianfei.zhang@intel.com>
Reviewed-by: Ilpo Järvinen <ilpo.jarvinen@linux.intel.com>